+++ users/cobis/parser.js
@@ -1,5 +1,4 @@
 // Requirements
-const fuzz           = require('fuzzball');
 const Option         = require('../../option').Option;
 const Author         = require('../../author').Author;
 
@@ -44,101 +43,20 @@ function parseAuthorOptions(author, bodies, callback) {
     let wikidataBody = bodies[0];
     let viafBody = bodies[1];
 
-    // Parse wikidata body
-    let knownViaf = [];
-    let parseCounter = 0;
-    parseWikidataOptions(wikidataBody, knownViaf, (wikidataOptions) => {
-        parseViafOptions(viafBody, knownViaf, (viafOptions) => {
-            // Compose options object
-            let options = wikidataOptions.concat(viafOptions);
-            // Get viaf details
+    // Get wikidata options
+    let wikidataOptions = parseWikidataOptions(wikidataBody);
+    let viafOptions = parseViafOptions(viafBody, wikidataOptions.filter(el => el.viaf).map(el => el.getViafId()));
+    let options = wikidataOptions.concat(viafOptions);
 
-                if(options.length > 0) {
-                enrichAllViafOptions(options, 0, function (options) {
-                    getAuthorSimilarOptions(author, options, (result) => {
-                        // Sort result
-                        result = result.sort((a, b) => (b.optionSuggested - a.optionSuggested));
-                        // Callback
-                        callback({'options': result, 'fields': authorFields});
-                    });
-                });
-            } else
-                // Callback
-                callback({'options': [], 'fields': authorFields});
-        });
+    // Enrich all options with VIAF and return them
+    Promise.all(options.map(el => el.enrichObjectWithViaf())).then(() => {
+        options.map(el => el.getString());
+        callback(options);
     });
 
 }
 
-
-function enrichAllViafOptions(options, position, callback) {
-    position++;
-    if (position === options.length) {
-        callback(options)
-
-    } else {
-
-        let option = options[position];
-        if (option.optionViaf) {
-            let viafUriParameters = option.optionViaf.split('/');
-            getViafDetails(viafUriParameters[viafUriParameters.length - 1], (result) => {
-
-                // Store titles
-                if(result.optionTitles) {
-                    // Generate object
-                    let optionTitlesObject = {'titlesSource': 'VIAF', 'titlesItem': result.optionTitles};
-                    // Handle Wikidata titles
-                    if(!option.optionTitles)
-                        option.optionTitles = [optionTitlesObject];
-                    else
-                        option.optionTitles.push(optionTitlesObject);
-                }
-
-                // Store occupations
-                if(result.optionOccupations) {
-                    // Merge occupations into description
-                    if(option.optionDescription)
-                        result.optionOccupations.unshift(option.optionDescription);
-                    option.optionDescription = result.optionOccupations.join(', ');
-                }
-
-                // Store birthDate
-                if (!option.optionBirthDate && result.optionBirthDate !== '0')
-                    option.optionBirthDate = result.optionBirthDate;
-
-                // Store deathDate
-                if (!option.optionDeathDate && result.optionDeathDate !== '0')
-                    option.optionDeathDate = result.optionDeathDate;
-
-                enrichAllViafOptions(options, position, callback)
-            });
-        } else {
-            enrichAllViafOptions(options, position, callback)
-        }
-    }
-}
-
-function parseWikidataOptions(wikidataBody, knownViaf, callback) {
-
-    // Wikidata map
-    let wikidataMap = {
-        'optionWikidata': 'wikidata',
-        'optionName': 'nome',
-        'optionType': 'tipologia',
-        'optionDescription': 'descrizione',
-        'optionTitles': 'titles',
-        'optionBirthDate': 'birthDate',
-        'optionDeathDate': 'deathDate',
-        'optionImage': 'immagine',
-        'optionWikipediaIt': 'itwikipedia',
-        'optionTreccani': 'treccani',
-        'optionViaf': 'viafurl',
-        'optionSbn': 'sbn',
-        'optionSuggested': null
-    };
-
-    // Results array
-    let wikidataOptions = [];
+function parseWikidataOptions(body) {
 
     // Parse results
     let results = body.results.bindings;
@@ -154,57 +72,46 @@ function parseViafOptions(body, viafUris) {
     let invalidFields = ['uniformtitleexpression', 'uniformtitlework'];
 
     // Parse results
-    let results = viafBody.result;
-    if(results && results !== null ) {
-        results.forEach((result) => {
-            // Generate new option
-            let viafOption = {};
-            // Check VIAF id
-            if (!knownViaf.includes(result['viafid']) && !invalidFields.includes(result['nametype'])) {
-                // Map result
-                Object.keys(viafMap).forEach((key) => {
-                    if (viafMap[key] && result[viafMap[key]] && result[viafMap[key]] !== '') {
-                        viafOption[key] = result[viafMap[key]];
-                        if(key === 'optionViaf') {
-                            knownViaf.push(viafOption[key]);
-                            // Get titles for option
-                            viafOption[key] = 'http://viaf.org/viaf/' + viafOption[key];
-                        }
-                        if(key === 'optionSbn')
-                            viafOption[key] = "IT_ICCU_" + viafOption[key].substring(0, 4).toUpperCase() + "_" + viafOption[key].substring(4, 10);
-                    } else
-                        viafOption[key] = null;
-                });
-                // Parse option for selection
-                viafOption['optionItem'] = JSON.stringify(viafOption);
-                // Push option
-                viafOptions.push(viafOption);
-            }
-        });
-    }
+    let results = body.result || [];
+    // Filter current results removing known authors and options with invalid fields
+    results = results.filter(el => !viafUris.includes(el['viafid']) && !invalidFields.includes(el['nametype']));
 
     // Construct options from query results
     return results.map(el => new Option(el, 'viaf', config));
 
 }
 
-function suggestOption(author, option){
+function parseOutput(data){
 
-    // Parse all options
-    if(author.titles && option.titles) {
+    let output = {};
+    let outputDictionary = config.getInputDictionary();
+    let configFields = config.getConfig().fields;
 
-        // Threshold 0.8 match result
-        let isSuggested = author.titles
-            .map(title => fuzz.extract(title, option.titles, {scorer: fuzz.token_set_ratio, cutoff: 80}))
-            .some(result => !!result.length);
+    // Translate data output to Beweb dictionary
+    Object.keys(data).map((field) => {
+        if((configFields[field].limit === null || configFields[field].limit > 1) && !Array.isArray(data[field]))
+            data[field] = [data[field]];
+        output[outputDictionary[field]] = data[field];
+    });
 
-        if(isSuggested)
-            // Set option as suggested
-            option.setOptionsAsSuggested()
+    // Get composite fields
+    Object.keys(configFields).filter(field => configFields[field].composite).forEach(compositeField => {
+        // Initialize output composite object
+        output[outputDictionary[compositeField]] = [];
+        // Populate output composite object and remove unnecessary composite fields
+        Object.keys(data).filter(field => field.includes(compositeField)).map(field => {
+            if((configFields[field].limit === null || configFields[field].limit > 1) && !Array.isArray(data[field]))
+                data[field] = [data[field]];
+            let subfieldKey = field.replace(compositeField, '').toLowerCase();
+            output[outputDictionary[compositeField]].push({[subfieldKey]: data[field]});
+            delete output[outputDictionary[field]];
+        });
+    });
 
-    }
+    return output;
 
 }
+
 
 // Exports
 exports.parseAuthor = (body) => {
@@ -213,6 +120,10 @@ exports.parseAuthor = (body) => {
 
 exports.parseAuthorOptions = (author, bodies, callback) => {
     parseAuthorOptions(author, bodies, callback);
+};
+
+exports.parseOutput = (data) => {
+    return parseOutput(data);
 };
 
 exports.configInit = (configObj) => {
